Make test method clickAdminMenuItem() private (other tests should use helpers, not hard-code menu names)
gotoAdminConsole() -> goToAdminConsole() for consistency with all other helper methods

Add keyword candidates for PostgreSQL 9.2

+++ api/src/org/labkey/api/collections/CaseInsensitiveMapTest.java
@@ -127,7 +127,7 @@ public class CaseInsensitiveMapTest extends Assert
 
         for (Future<Integer> future : futures)
         {
-            assertEquals(0, (int)future.get());
+            assertEquals(0, future.get());
         }
     }
 }
